requirements.txt
astroid==2.0.4
autopep8==1.4.1
certifi==2018.10.15
chardet==3.0.4
colorclass==2.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
coveralls==1.5.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.2
django-cors-middleware==1.3.1
django-extensions==2.1.3
django-filter==2.0.0
django-heroku==0.3.1
django-model-utils==3.1.2
django-nose==1.4.6
django-notifications-hq==1.5.0
django-password-reset==2.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
docopt==0.6.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
oauthlib==2.1.0
openapi-codec==1.3.2
packaging==18.0
Pillow==5.3.0
pip-review==1.0
pip-upgrader==1.4.6
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.1.1
pyparsing==2.2.2
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
sendgrid==5.6.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==3.0.0
social-auth-core==2.0.0
terminaltables==3.1.0
typing==3.6.6
uritemplate==3.0.0
urllib3==1.24
virtualenv==16.0.0
whitenoise==4.1
wrapt==1.10.11
